requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.15.1
channels==3.0.4
channels-redis==4.1.0
charset-normalizer==3.3.0
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.1
daphne==3.0.2
Django==4.0
django-extensions==3.2.1
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
huggingface-hub==0.18.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
multidict==6.0.4
packaging==23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyOpenSSL==23.2.0
pytz==2023.3
PyYAML==6.0.1
redis==4.5.5
requests==2.31.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
text-generation==0.6.1
tqdm==4.66.1
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.7
yarl==1.9.2
zope.interface==6.0
